[cherrypick] Point jsonrpsee to forked version to support error metrics and use camelCase for unwrappedThenDeleted (#12147)
## Description 

1. Point jsonrpsee to forked version to support error metrics by error
code
2. [bug fix] previously, effects had an unwrapped_then_deleted field on
ts-sdk. This is an issue since jsonrpc returns the field as
unwrappedThenDeleted, and caused errors on making certain get
transaction block requests.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes