indexer: update TPS calculation to be checkpoint based (#12355)
## Description 

Before this change, the TPS was calculated over time window of 10
seconds, which was not able to catch spike of one checkpoint with many
tx.

Changing this to be checkpoint based, which will catch a spike
checkpoint and also the max consistent TPS we can get.

More details can be found in thread:
https://mysten-labs.slack.com/archives/C032676BWGN/p1685996135421019?thread_ts=1685996115.167529&cid=C032676BWGN

## Test Plan 

This change has been on production mainnet. 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes