[sui-execution] CutPlan::execute
Implement the process of executing a cut plan to copy packages and
fix-up manifest files (for the copied crates, and the workspace).

Also implements `CutPlan::rollback` which is used by `execute` to
clean-up after itself, if it fails.  To support this,
`CutPlan::discover` has been updated to track the new directories it
needed to create.

This PR does not address various usability issues (adding context to
error messages, a dry-run option, etc) which are to come.

Test Plan:

New unit tests:
```
sui-execution/cut$ cargo nextest run
```

Try the CLI on a real example (warning, this will make changes to the
repo's working tree):

```
sui$ cargo run --bin cut -- --feature v1                 \
  -d sui-execution/latest:sui-execution/v1:-latest       \
  -d external-crates/move:external-crates/move-execution \
  -p sui-adapter-latest                                  \
  -p sui-move-natives-latest                             \
  -p sui-verifier-latest                                 \
  -p move-bytecode-verifier                              \
  -p move-stdlib                                         \
  -p move-vm-runtime
```