[bcs] Fix a bcs decoding bug for u128 and u256 values (#12474)
## Description 

fixes #12470

When parsing u128 and u256 values we were incorrectly padding the lower
bits. This resulted in incorrectly decoding values in the following
ranges:
* u128 values with bytes 9-16 being 0x00
* u256 values with bytes 17-32 all being 0x00
* u256 values with bytes 9-16 all being 0x00

## Test Plan 

Updated a couple test cases that triggered this issue

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes