[move-unit-test] Add stack traces for tests that abort in natives (#12472)
## Description 

Fixes an issue where Move unit tests that had an unexpected abort in a
native function would not report a stack trace since we didn't capture
execution state on a call failure to a native function.

## Test Plan 

Added a regression test to make sure that we produce stack traces for
native functions that abort.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
* Adds stack traces for failing Move unit tests that abort in a Move
native function.