[explorer] Add object video support (#12491)
## Description 

This adds video support to object displays in explorer. Some notes:
- The object display must have `video_url` property, _in addition_ to an
`image_url` field. The `image_url` is expected to be the thumbnail of
the video, and the fallback display in the event that videos are not
supported. This could change in the future, but for now seems like a
sensible approach.
- Only supported for recognized packages.

## Test Plan 

Ran locally.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes