Cherry pick unwrapped then deleted simplification to release branch (#12600)
## Description 

This PR cherry-picks the change to enable
simplified_unwrapped_then_delete.

## Test Plan 

CI + devnet

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Prior to this change, unwrapped_then_deleted field in TransactionEffects
only contain unwrapped then deleted objects that previously existed in
the store (i.e. was once unwrapped in its lifetime). If an object was
always wrapped since creation and was never unwrapped ever, it wouldn't
show up in unwrapped_then_deleted. This this change, we no longer makes
such distinction. As long as an object is unwrapped then deleted in a
transaction, it would show up in effects regardless of its history. This
simplifies the logic of handling them significantly.
To properly maintain the state of the system accurately with this
change, at the epoch boundary when this upgrades happens, we must
re-accumulate the state root hash of all objects in the store. This
process takes a few seconds on mainnet. Testnet can take a much longer
time (30s-1min) since it contains significantly more objects. Users may
observe a brief slow down of the network during epoch change when this
is applied.