move: source service config (#12552)
## Description 

This adds a basic config file to initialize the server with. It takes a
list of `Packages`, where `Packages` comprise a repository and one or
more paths to Move package in that repository.

## Test Plan 

Added a basic test to parse a config file.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes