Integrate state sync with archive (#12427)
## Description 

This PR adds the ability to read checkpoints from archive store if all
peers of a node have purged checkpoints that the node is looking for. In
particular, we add a state sync archival task which on noticing that for
each of its peer when condition `current_node.highest_synced_checkpoint
< peer.lowest_available_checkpoint` is true, downloads the missing
checkpoints from the archive.

## Test Plan 

Added a test which sets up two nodes where one of the node prunes older
checkpoints. And archive is setup with older data beforehand. A new node
is expected to have downloaded all data i.e. pruned (from archive) and
latest from peers.