[sui-execution] Cut v0 (#12698)
## Description 

Add a script to cut a new version of `sui-execution` and
`move-execution` (based on the `sui-execution-cut` crate), and use it to
cut a v0. Also introduces protocol version 17 to switch from using `v0`
to using `latest` (both are identical as of this commit).

Note that `v0` is **identical** to `latest` at the time of this PR --
there will be follow-up changes to `latest` that can take advantage of
its decoupling with the implementation used in previous versions.

## Test Plan 

Run all tests:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```

- Once with all the changes except the protocol version change,
- Once including the protocol change

(To test both cuts).

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Introduces Protocol Version 17, which includes a new versioning scheme
for the execution layer.