[execution][dynamic fields] Support versioned child object reads (#12055)
## Description 

- We now track wrapped UIDs when loading an object in execution
- This information is used to build a map from UID to the root object's
version at the beginning of the transaction
- We then bound a child object version when doing dynamic field lookups

## Test Plan 

- We only have 1 existing test that tries reading a dynamic field with
an "old" parent version, but that test is for dev inspect.
- Added tests that leverage dev-inspect to load old versions of objects 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

This release fixes dev-inspect's handling of old versions of dynamic
fields. Prior to this change, the runtime always fetched the most recent
version of dynamic fields. For example if the most recent version of an
object was version `10`, and you used dev-inspect to query a dynamic
field on version `7` of the object, you would still see the dynamic
field for version `10`. After this change, you will correctly view the
dynamic field as it was on version `7`.