Improve db checkpoint handling (#12773)
## Description 

The current workflow of uploading db checkpoint only uploads in sequence
i.e. it would wait forever if there could be gaps in epoch snapshot on
the remote storage. This is not a requirement and more often, we really
want to upload latest epochs and don't care if older epoch snapshots are
there or not.
This PR now makes it such that if there is a db checkpoint which is
present local but not on remote, it will upload it. Besides, it also
improves how garbage collection runs by making it not dependent on
upload.

## Test Plan 

Added uni tests