fix: add protocol version for upgraded multisig (#12812)
## Description 

1.4.0 accepts both legacy and upgraded multisig. 1.3.0 only accepts
legacy multisig. this is to make upgraded multisig enabled only for
protocol version >= 18 when all validators upgrade.
## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes