[sui-execution] Test to make sure generated lib.rs is not changed
Add a step to the rust `test` workflow to regenerate `lib.rs`, so that
if the PR has inadvertantly introduced a manual change to this module,
the `change-files` step will identify that and complain.

Test Plan:

Lints and formatters:

```
$ flake8 scripts/execution_layer.py
$ black scripts/execution_layer.py
```

Diff the generated (`--dy-run`) contents before and after generating
`lib.rs` for real.  If there was a difference before, it should be
gone after generation runs for real:

```
$ diff -u ./sui-execution/src/lib.rs \
  <(./scripts/execution_layer.py --dry-run generate-lib)
$ ./scripts/execution_layer.py generate-lib
$ diff -u ./sui-execution/src/lib.rs \
  <(./scripts/execution_layer.py --dry-run generate-lib)
```

Check in CI that the step runs and catches a discrepency -- fix the
discrepency and notice CI finishes successfully.