[Deepbook] Market ask order with quantity #12819 ## Description
Update market ask order with quantity. Currently the quantity argument
in place_market_order only used for bid side and we sell all the passed in
base asset instead the quantity of base asset. To avoid confusing, we now
also sell the quantity of base asset.

## Test Plan

cd to deepbook dir and run "sui move test"

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [✅] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [✅] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

---------