cherrypick [12861]: refactor streaming code here and there (#12861) (#13057)
## Description 

Cherry pick https://github.com/MystenLabs/sui/pull/12861

Refactor the streaming code in the following way:
1. change streaming from blocking (`send`) to unblocking (`try_send`).
If the channel is full because consumer is slow for example, err and
drop the subscription. This mitigates the fullnode slowness issue when
the subscription channels are full and too many tokio tasks are spawned
and pending.
2. this also makes the items sent in strict order
3. do not copy the subscriber maps for every item. Instead hold the read
lock for a short period of time
4. remove bad subscriptions in the end to minimize the duration & times
we hold onto the write lock.
5. add more metrics, and separate metrics for event streaming and tx
streaming
6. rename "event_handler.rs" to "subscription_handler.rs"

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Refactor the streaming code for stronger robustness, higher efficiency
and in-order guarantee.

## Description 

Describe the changes or additions included in this PR.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Co-authored-by: Lu Zhang <8418040+longbowlu@users.noreply.github.com>