sdk: fix personal message to use bcs (#12997)
## Description 

fix a bug that the signMessage in rust vs typescript has a discrepancy.
this is make typescript use bcs serialization for a vector instead of
just a plain vector.

```
PersonalMessage {
  message: Vec<u8>
}
```
## Test Plan 

existing tests in ts still passes (bc verify and sign are both changed
to the new format).
tested against rust that a ts produced sig now verifies in rust. 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: Michael Hayes <michael.hayes@mystenlabs.com>