update all apps to use modular imports from ts-sdk (#13070)
## Description 

This updates imports from the SDK to use the new modular exports across
all apps in the repo.

The biggest change (other than the specific imports) is that the
JsonRpcProvider is being replaced with SuiClient. This expose mostly the
same API, but the setup is slightly different, and the SuiClient doesn't
support faucets.

## Test Plan 

Asking for help to do some manual testing from the team

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes