[move-compiler] Added coin field linter (#13144)
## Description 

Added a linter signaling that a struct field contains a field of type
`sui::coin::Coin` and that using a field of type `sui::balance::Balance`
instead would be more efficient.

Cleaned up the linting code a little as well.

## Test Plan 

A test has been added.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

When building Move code, additional linter warnings related to structs
containing fields of type `sui::coin::Coin` may appear. The goal of this
linter is to encourage developers to use fields of type
`sui::balance::Balance` instead as this solution is more
space-efficient.