A few simplifications in the sui-executor (#13140)
## Description 

1. In the circular ownership check, input objects cannot have object
owner, so they don't need to be involved at all. Remove them. @tnowacki
please double check if this is correct.
2. Rename ObjectStore to ChildObjectStore to improve accuracy
3. Fix a few indentations

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes