[EffectsV2] Lift up a few effects APIs (#13055)
## Description 

There are a few API functions in TransactionEffects that can be derived
from other functions. They don't need to be defined per-effect type, but
can be defined at top level. This will reduce code duplication when we
introduce new effects variants.
Also change the semantics of all_deleted to exclude
unwrapped_then_deleted objects and rename it, because
unwrapped_then_deleted are never needed with this API call.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes