[deepbook] adding rounding on balance locked (#13213)
When the `order_price` and `order_quantity` are small enough such that
there is a rounding down of
the balance, a custodian trying to cancel their limit order will leave a
single MIST in their `locked_balances`.

So we need to round up in order to fully remove the locked balance from
the account.

How did you test the new or updated feature?

unit test
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [x] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration