indexer fix: split commits of obj mutation and deletion (#13236)
this is a bug but did not manifest when we committed each checkpoint one
by one.
after recent change of writing multiple checkpoints together, an error
of ON CONFLICT DO UPDATE would happen.
specifically, if in an object commit batch, an object is mutated first
and then deleted later, this error will show up.
thanks to @bschwab for detecting and reporting this.

test locally and with bryan on the snowflake pipeline.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration