[wallet-ext] - update coin balance / tokens UI (#13259)
## Description 


Updates the `TokensDetails` page to match the latest designs. That whole
file needs a lot of love, but decided to work with it this time 🤝



![image](https://github.com/MystenLabs/sui/assets/122397493/2c37e246-f74a-4890-9793-ab1bd1166ff9)

**no SUI (testnet)**

![image](https://github.com/MystenLabs/sui/assets/122397493/c64fdbb0-21d2-407e-acc6-f8fc3c8b01d1)

**no SUI (mainnet)**

![image](https://github.com/MystenLabs/sui/assets/122397493/702485fc-b9b5-4fef-bae1-da0e50f7f8f1)






## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes