move: field to watch upgraded objects in source service (#13376)
## Description 

Adds ability to specify `watch = <Object ID>` to track on-chain
upgrades. This PR is scaffolding and only adds the option to the config
and the Rust types, no actual tracking of the `Object ID`.

## Test Plan 

Updated tests for config schema.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes