[move-compiler][sui-mode] Add init and one-time witness rules (#13335)
## Description 

- Added init and one-time witness rules to the compiler's Sui mode

## Test Plan 

- Migrated tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

In this release, the Move compiler will now error on incorrect
declarations (or usage) of `init` functions and one-time witnesses.