[Deepbook][1/2] Migrate TS SDK to V2 + Re-write (#13273)
## Description 

This is a re-write for https://github.com/MystenLabs/sui/pull/13192.
Changes include:
- Fixed build & linter errors and styling inconsistencies(e.g., snake
cases)
- Added E2E tests to verify the correctness of implementation and make
sure the SDK stay up to date with the onchain code
- Refactoring the code. The previous code is divided into
`Deepbook_query` and `Deepbook_sdk`. After the refactoring it would be
`DeepbookClient` and `Router`(`Router` code will be re-added in the
following PR)
- Simplify the API by removing parameters that can be derived
- Decode values returned by devinspect for functions like
`listOpenOrders`, `getMarketPrice` such that the user will get the
actual values instead of BCS bytes
- Add new util functions including `getAllPools`

## Test Plan 

Added end to end tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [x] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: 0xdeadbeef <18015565550@189.cn>