Refactor check_sui_conserved (#13440)
## Description 

This PR makes two changes:
1. Refactor `check_sui_conserved`. This function had a lot of redundant
code, where it iterates through writes and deletes, and for each of them
it operates on the input and output. Instead of duplicating code, this
PR introduces a function that returns the list of inputs/outputs for the
function to operate on. The operation then becomes much simpler.
2. Instead of reading DB for input objects, this PR reads from the
metadata that's already in the temporary store, avoiding unnecessary DB
reads.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes