basic distributed execution without proper sharding
Sharding of the objects store is not yet implemented. The current implementation of distributed execution assumes that one EW executes all txs modifying the system object (0x5). For this to be compatible with state sharding, we would need a mechanism for (ideally locally) figuring out whether a given object is a child object (at some level) of the system state.