Remove unnecessary LocalProtocolConfig clone. Rename LocalProtocolConfig for style concistency. Add format logging. (#13554)
## Description 

LocalProtocolConfig was needlessly cloned. 
Additionally, renaming LocalProtocolConfig variable to bring it inline.
I thought it was a bit funky that it was named constants when it was a
bit more than that.

## Test Plan 

- none

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes