treat ObjectNotFound and DependentPackageNotFound as non-retryable (#13608)
## Description 

In authority aggregator we treat ObjectNotFound and
DependentPackageNotFound as retryable, but they have special treatment
than other retryable errors. When we return the error to users, whether
these two types of error is retryable depends. In this PR we choose a
safer route, always treat them as non-retryable. We also sort the errors
by total stake, to make clients see the most prominent errors first.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
In RPC response processing, treat ObjectNotFound and
DependentPackageNotFound as non-retryable