Analtics pipeline (#13511)
## Description 

Initial implementation of the analytics pipeline. 
I am thinking it would be best to start reviewing this and committing
sooner as opposed to get this in a great shape and put it up for review
only then. It's getting bigger and there is still quite a bit to do and
it will only get bigger and bigger, so we better start moving.
The overall idea is here though. So please take a look and let me know
if you see issues that you want solved or done differently, or just
broken.
@sadhansood, @bschwab and I talked about this and trying to get it in
seems important, so others can work on it as well, and we can iterate.
Nothing uses it for now. Last I tried it, on my machine, with a local
fullnode, it collect checkpoints and runs. So it's usable.
Next step is to write a csv writer and start the task that reads those
files. Sadhan is implementing the uploader to snowflake.

Schema are likely to change but that is ok.

## Test Plan 

Nothing for now and we need to come up with something, not sure what
though. Suggestion welcome.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes