[framework-snapshots] Don't assume there's a max_ver - 1 snapshot
The test that upgrading from the version of the package currently
running on a production network to the max version in `main`
previously assumed that there would always be a snapshot available for
`max_ver - 1`.  This may not be the case, if a release goes out with
multiple protocol version bumps.

All other tests that reference snapshots do so via the snapshot
manifest, so this change updates the upgrade compatibility test to do
the same: It finds the largest version strictly less than `max_ver`
that we have a snapshot for.

Test Plan:

```
sui simtest test_upgrade_compatibility
```