[dapp-kit] add new useConnectWallet mutation hook (#13661)
## Description 

This PR adds a new `useConnectWallet` mutation hook with tests for
connecting to a wallet. I was going with the idea that we should create
these mutation hooks as more of barebone wrappers on top of react-query
so folks can still provide customization options along with things like
mutation keys & use interfaces that they are familiar with (i.e., we
return mutate instead of connectWallet). The whole mutation/query key
story might need a little more thought, but I think I made a reasonable
attempt by exposing the mutation key factory. Glad to hear feedback on
that!

Equivalent PR for disconnecting from your wallet connection:
https://github.com/MystenLabs/sui/pull/13706


## Test Plan 
- Automated tests with pretty thorough coverage
- Tested that the functionality worked in an example dApp for a variety
of scenarios
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes