[graphql][1/n] Refactor gas-related convert_to fns to utilize async-graphql's Object macro (#13714)
## Description 

This PR refactors convert_to_gas_input, convert_to_gas_cost_summary, and
convert_to_gas_effects in favor of utilizing the `#[Object]` macro.
These convert_to functions are intended to convert struct objects into
GraphQL objects, but the issue is that they will always make
computations and data fetches for each field, even if the caller did not
specify those fields. For example, if I make a query for
TransactionBlock for the gasInput field and only specify the gas_price,
we nonetheless make the extra conversions for gas_sponsor, gas_budget,
and the data fetch for gas_payment.

Continuing with the example above, in this refactor, the struct
definition of the GraphQL object GasInput now has the corresponding
"native" fields, and implements From SuiGasData. The impl Object for
GasInput block then resolves those fields, doing the necessary
conversions and data fetches.

1/n as there are other convert_to fns I'd like to refactor.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes