[dapp-kit] add new useSwitchAccount mutation hook (#13753)
## Description 

This PR adds a new `useSwitchAccount` mutation hook for switching
between different wallet accounts. Not sure if we prefer
`useSelectAccount` or `useSwitchAccount` for naming, but I thought
saying "I'm using this hook to switch accounts" out loud made more sense
than "I'm using this hook to select an account" 😛 .

## Test Plan 
- Automated tests
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes