[dapp-kit] update wallet state based on properties such as accounts changing (#13748)
## Description 

Some properties of a wallet might change during the course of
interaction with a dApp; accounts might be deleted/added or wallet
features might be updated. One realistic example here is where a user
connects some number of accounts to a dApp and then removes their
actively connected account from their wallet. In this case, we would
want to just default the user's current account to the first connected
account they have.

## Test Plan 
- Automated tests
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes