[packages] Fix dependency override prunning issue (#13921)
## Description 

Fixes an issue related to pruning overridden dependencies during
dependency graph construction. More concretely, when tryin to find
dependency paths in a pruned sub-graph, one could encounter no target
node if this node was representing a pruned overridden package. The
proposed solution is to simply also search for the target node in the
overrides map (if a search for it in the sub-graph fails)

## Test Plan 

Added a test for this specific scenario, plus all existing tests must
pass.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes