[2/n][transfer-to-object] Add pruning of object marker table at epoch boundaries (#13420)
## Description 

All elements of the `object_per_epoch_marker` table can be pruned at
epoch boundaries, so this adds this in to the reconfiguration step that
transitions between each epoch. I believe the way I have it here is the
correct way, but I'm less familiar with this so please let me know if
there's a better way to do this.

## Test Plan 

Added e2e tests with reconfiguration enabled, and manually checked to
make sure any entries were cleared out at epoch boundaries.