[GraphQL] BigInt from primitive integer types
Generate `From<uint>` implementations for `BigInt` for each of the
unsigned integer types.  This is necessary to support the representing
Move Values which are stored in a primitive form, and need to be
converted to a `BigInt`.

Also makes `BigInt` serializable, with a transparent representation,
for use within the `MoveValue` type.

Test Plan:

Tests have been simplified to be written in terms of `BigInt`'s
`FromStr` implementation, which allows us to avoid having to test both
the `InputType` and `ScalarType` paths.

Additional tests have been added for conversion from primitive types.

```
sui-graphql-rpc$ cargo nextest run -- big_int
```