[GraphQL] Support for MoveValue
Adds a new GraphQL object type: `MoveValue` and a new scalar type to
represent structured move value contents: `MoveData`.

This PR proposes that `MoveData` should be a scalar rather than a
GraphQL (output) type because of its recursive nature:  It would be
difficult/impossible to selectively query into its nested structure.

Test Plan:

This type and scalar are not currently hooked into the schema, because
to do so requires implementing a module resolver that can be used to
generate a type layout for any given type, but unit tests have been
added for the value representation:

```
sui-graphql-rpc$ cargo nextest run -- move_value
```