[EffectsV2 3/n] Refactor TemporaryStore to be ready for effects V2 (#13560)
## Description 

This PR refactors TemporaryStore so that it could emit both effects v1
and v2.
Previously, TemporaryStore contains a `written` field and a `deleted`
field that contains all the writes and deletions, tailored for effects
v1.
This PR changes the output field of TemporaryStore to use
ExecutionResultsV2 directly. ExecutionResultsV2 contains all the
primitive data of the execution output.
In order to maintain all the primitive results properly, this PR
replaced the `write_object` and `delete_object` function with a list of
more specific APIs such as `mutate_input_object`, `mutate_child_object`,
`delete_input_object`, `create_object`, `update_system_package` and etc.
At the end, when we emit effects, we generate the object changes based
on the primitive results, and convert back to what effects V1 needed.

## Test Plan 

CI.
Also replay from genesis on mainnet.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes