[dapp-kit] theme the UI components 🎨  (#13994)
## Description 

This PR builds on the theme definition PR to fix up our styling and to
actually leverage the theme definitions. I'd say this is looking _pretty
good_ - I think really we need to test out some themes / try porting
this into SuiFrens, and customizing the connect button / dropdown to
make sure we're providing the right level of customization. There might
also be a little more design polish work as well for things like the
"What is a wallet button" on mobile.

Follow-ups to get this a bit better shape since I'm out on PTO the rest
of the week:
- There's one FIXME to align the back and X icons because they don't
align on mobile
- The "What is a Wallet" button could use some love on the design side
- There's a CI failure for PostCSS even though I didn't change anything
and rebased on main
- We might need to add some more theme variables for things like the
border between the left & right modal panels


https://github.com/MystenLabs/sui/assets/7453188/a5da40d6-3d1f-45e8-9348-3c9ecb41d7e6

## Test Plan 
- Manual testing

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: Jordan Gensler <jordan@mystenlabs.com>