[on-chain-deps] use original package names in resolution error messages
When using pkg name resolution, dependency error messages during graph
resolution will print resolved names instead of the original ones as
defined in manifests which will be confusing to the user. This commit
makes it so that original package names are used. This is achieved by
storing the original name in the `Dependency` struct which is the edge
between two packages in the graph. This way the original name can
always be recovered.