[graphql/rpc] transaction block kind (#14129)
## Description 

Add `TransactionBlockKind` resolvers in Graphql. Not fully completed due
to missing other types (Move stuff), which we will add later. 

Changed the Graphql schema for GenesisTransaction to take as a field a vector of SuiAddresses, which matches what we have in JSON-RPC today.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes