[on-chain-deps] implement sui package hooks
Implements `resolve_on_chain_dependency`, `custom_resolve_pkg_name`, and
`resolve_version` hooks for Sui so that on-chain dependencies are
supported.

This is achieved by resolving dependency names to their `original_id`
(the id if package version 1) and their version to their on-chain
version when the `published-at` field is present in their manifest.
In effect, this consolidates the dependency graph so that on-chain and
source transitive dependencies can get deduplicated by changing the
names of source dependencies to their `original_id` and assigning them
their on-chain version.

On-chain dependencies are resolved by fetching their contents from the
RPC to their respective location in `~/.move/` directory. RPC from
current active environment is used.