[on-chain-deps] add addresses for system packages to generated manifest
This alleviates a bit the problem where the dependency graph picks an on-chain
package as dependency for a source package (instead of source) when both
are present in the graph which causes undefined address name errors.
It does not solve the problem for all packages, but in the worst case,
user can use an override until a proper solution is worked out.