[on-chain-deps] exclude on-chain dependencies from source verification
Move core's `CompiledPackage` doesn't include bytecode dependencies in
the result. This causes two problems in dependency verification.

First is that the topo sort in `get_dependency_sorted_modules` panics
because it relies on building a dependency graph first which fails due
to missing dependencies. And second is that dependency verification
tries to verify on-chain dependencies also (which is unnecesssary as
they were fetched from the chain anyways).

This is solved by doing topo sort directly without relying on building a
dependency graph and excluding on-chain deps from verification.