[sui-benchmark] run multiple traffic profiles on same benchmark (#14054)
## Description 

Modifying the sui-benchmark suite in order to allow us to run benchmarks
with multiple traffic profiles instead of only one constant rate as we
currently do. That will allow us to simulate load of variable traffic
(and type of txes etc) instead of constant one. To achieve this the
`Bench` cli subcommand was extended to provide all the arguments as
vectors allowing us to set multiple "benchmark groups" , where the
properties of each benchmark group are all the values of the specified
vector position. More information about this on the `options.rs` file.

To achieve the above we build workloads for each "benchmark group" and
then we initialise all those once. Then a bench workers schedule task
has been introduced to coordinate running each benchmark group. Each
group is defined to run for a specified `duration`. That allow us to do
things like `run with 1K tx/s for 100 seconds, then run with 500 tx/s
for 80seconds...` . Similarly we can configure further the groups to run
a benchmark like `run with 1K tx/s for 100 seconds with 50% shared & 50%
owned transactions , then run with 500 tx/s for 80 seconds with 80% own
transactions and 20% adversarial transactions`.

Also, the benchmark groups will run in _a round robin fashion_. For
example, if we do configure a benchmark with 3 groups, then after the
last group is run, it will go again and start from the beginning - first
group. That allow us with minimal configuration (just defining a few
groups) to run repetitive traffic patterns and test the system's
behaviour. If a group is defined with duration `unbounded` then that
will run for the rest of the benchmark (so someone needs to be careful
with the setup). From a software perspective the `BenchWorker`s created
for each group are re-used during the round robin so no need to do any
re-setup while running or spend excessive resources . The only thing
that is managed is the tasks that are spawn and complete.

**Backwards compatibility:** the cli parameter changes are backwards
compatible so existing CI jobs won't break. Existing CI jobs will
basically run with default settings having only one benchmark group.

## Test Plan 

An example of running a benchmark test that defines 2 benchmark groups
where we want to test spiked traffic:
* 1000 tx/sec for 360sec (100% shared objects)
* 2000 tx/sec for 120sec (100% shared objects)

the **bench** sub-command line parameters look like:
```
--num-of-benchmark-groups 2 \
--in-flight-ratio 30,30 \
--num-workers 24,32 \
--target-qps 1000,2000 \
--shared-counter 100,100 \
--transfer-object 0,0 \
--delegation 0,0 \
--shared-counter-hotness-factor 50,50 \
--batch-payment 0,0 \
--batch-payment-size 100,100 \
--adversarial 0,0
--adversarial-cfg 0-1.0,0-1.0 \
--shared-counter-max-tip 0,0 \
--duration 360s,120s
```

as we can see on the above parameters we can configure all the arguments
independently for each benchmark group and those will get respected on
each run.

on the graph bellow we can see the `input traffic to consensus` is
following the benchmark setup alternating between the 2 traffic profiles
for the respective defined durations:

<img width="716" alt="Screenshot 2023-10-03 at 00 28 59"
src="https://github.com/MystenLabs/sui/assets/17335598/86564a64-deac-4738-83b9-ffc7aeea6075">


---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes