[dapp-kit] do a little polishing on the hooks for inferring the active chain / exposing errors / etc. (#14181)
## Description 

This PR does a little polish on the signing hooks for a few things I
noticed were off. The most notable change here is that we're making the
`chain` arg optional as we should just infer the active chain from the
account. This keeps parity with the wallet-kit implementation

## Test Plan 
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes