Add new @mysten/create-dapp package (#14191)
## Description 

This adds a new `@mysten/create-dapp` package which can be used as:

```bash
pnpm create @mysten/create-dapp
```

This will prompt the user to select a template, and a name for their app
and then create all the files required to get started with a new dApp.


Currently there are 2 templates. one is a simple client only app that
connects to wallets and shows their owned objects. The other is a simple
e2e counter example. The templates still need some more documentation
that can be added in a future PR

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: William Robertson <williamrobertson13@users.noreply.github.com>