[move-compiler] revised hlir lowering (#13974)
## Description 

This revises how HLIR lowering proceeds to be more like typical context
normalization. In doing so, it elides a fair bit of code that was being
previously generation around effects in value position.

## Test Plan 

The tests mostly all work, and a few have been updated with additional
errors.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

Some new errors will be generated.

### Release notes