Fix a bug in full checkpoint content constructor (#14232)
## Description 

Somehow the full checkpoint content constructor is ignoring the user
signatures even though it has it.
This PR fixes it. I took a look around and my guess is it's copied from
a constructor in checkpoint content, which we should really rename it to
make it test-only.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes