[dapp-kit] add global connection status info to the useCurrentWallet hook along with a new useAutoConnectionStatus hook (#14284)
## Description 

As I was migrating SuiFrens over to dapp-kit, I noticed that there
really isn't a great way to get the global connection state of your
wallet in relation to the dApp. There are scenarios in SuiFrens where we
want to show a custom connect button with the built-in modal component
and render different UI elements based on the connection status, but we
currently have no access to this information. @Jordan-Mysten also ran
into a similar problem where he needed to determine whether or not an
auto-connection attempt was in progress.

To address this, I'm adding a react-query style `connectionStatus` with
derived booleans to the `useCurrentWallet` hook since that information
changes together. This is technically a breaking change hence the major
bump JFYI.


## Test Plan 
- Automated tests
- Manual testing
- Tested docs

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes