[GraphQL/Object] Tweak BCS implementation
The BCS output for Object should represent its raw bytes (what you
would see if you looked at it in a node's DB), which differs from its
current format which outputs the BCS representation of an object if
it's a `MoveObject`, or the BCS representation of a package if it's a
`MovePackage`.

This also has the unfortunate side effect of discarding the
information about whether the Object is a Package or a MoveObject
after it has been initialised.

This PR adapts the BCS representation for Object to be the
serialization of its on-chain representation (this can be done more
efficiently when we introduce data loaders and sunset the existing
interface because this is exactly the contents of one of the DB
fields).

In future PRs, I will introduce the `bcs` representation of a
`MovePackage` explicitly and connect up `MoveObject` so that it
becomes possible to access all possible BCS representations.

Test Plan:

Manually tested with GraphiQL.

<Screenshot TBD>