[cherry-pick] Make formal snapshot creation wait until db checkpoint is uploaded (#… (#14403)
…14214)

## Description 

Formal snapshot creation starts as soon as it notices a db checkpoint
directory. And tries to open the db to create a live object set
iterator. But because we also do pruning and compaction on the db
checkpoint before we upload, a different thread tries to open the db to
do pruning and compaction at the same time and panics. The fix is for
formal snapshot creation workflow to wait until db checkpoint is already
uploaded before it will consider it.

## Description 

Describe the changes or additions included in this PR.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes