[GraphQL/MovePackage] BCS field
Field to get the BCS representation of a package's modules.  The
decision to only output the modules, rather than the whole package
struct was done to match the format that modules take when they are
passed in to the network, during publish and upgrade transactions.

Test Plan:

Tested in GraphiQL:

<TODO: Screenshot>

And using cURL:

```
curl -LX POST "http://127.0.0.1:9001" \
    --data-raw '{
      "query": "{ object(address: \"0x2\") { asMovePackage { bcs } } }"
    }' \
  | jq  -r .data.object.asMovePackage.bcs
  | base64 -d
  | hexdump -C
```