[benchmark] Update Move workload (#14427)
## Description 

This PR updates the sui-single-node-benchmark's workload such that there
is a fair comparison between non-Move and Move workload. Specifically it
makes the num_input_objects a common parameter that can be specified in
both non-Move and Move mode.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes