[GraphQL/Object] asStake is allowed to fail
`as...` fields are special in that they are fallible conversion
operators: It's a valid query to request stake-specific information on
an object that isn't a `StakedSui`.

This PR changes the behaviour of the `asStake` to match this pattern.
It also tweaks `asCoin` to follow a similar coding pattern.

Test Plan:

Tested with GraphiQL:

<TODO: Screenshot>