Allow object downloads without authentication from certain remote object stores (#14388)
## Description 

We want to allow object download without authentication from certain
remote stores with free egress i.e. cloudflare, wasabi, etc. This PR
adds a new downloader module which allows us to do that for S3, GCS, etc
Next step would be to integrate it in sui tool and let users benefit
from it.

## Test Plan 

Added unit test. Also tested with S3 and found it to be working