[CI] Ensure Move Tests are run when Move files change
Move tests are run as part of the Rust tests for the Sui Framework,
but they should also be run when there are only Move changes.

Surfaced because a recent PR (#14429) Moved some Move tests causing CI
to fail on `main`, and this wasn't picked up because diff tests hadn't
been run.

Test Plan:

CI