[consensus_adapter] connect the consensus throughput profiler to the submission side (#14118)
## Description 

Connecting the throughput calculator to the submission side. The
throughput thresholds are calculated per chain based on the chain id.
According to each throughput profile a different delay step is assigned
for the submitter of `position = 1` . Also now 3 traffic profiles are
supported, Low, Medium & High. The latencies for each profile are set
based on observations of the mainnet environment to ensure that under
each load the desired amplification is maintained (as throughput
increases we want less amplification. For the case of `Low` traffic we
always make the validator of position = 1 co-submit, so basically a user
transaction will always get submitted by two validators on the same time
when traffic is low.

The goal is to load test in testnet before doing any rollout to mainnet.

## Test Plan 

CI & private-testnet

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes