[Move/Tests] Ensure examples build outside of test mode
A recent change to the toolchain caused the examples to fail to build,
because they `use`-d `test_scenario` without guarding the import with
`#[test_only]`.

Previously this seemed to work because imports were largely ignored
unless the imported value was referenced, and at that point, there may
be an unbound reference issue.

Now, the imports seem to be eagerly referenced, which causes a problem
because `test_scenario` is itself a `#[test_only]` module.

This PR introduces an extra step to the CI support for Move tests to
also try building example packages in non-test mode, to make sure that
continues to work.

This extra step has only been introduced for example packages, because
the system package builds are already covered by CI (they already need
to build successfully for us to build the node software, CLI, etc).

Test Plan:

```
sui-framework-tests$ cargo nextest run
```