[Docs/Examples] Purge References to `sui_programmability`
Examples in `sui_programmability` are out-of-date, and we should
prefer referencing examples in `examples/move` instead.  This PR gets
updates most references to `sui_programmability` in the docs, leaving
only two to be picked up in follow-ups:

- Flash Lender
- Shared vs Owned Objects

Where in both these cases more examples need to be modernised and
ported to `examples/move` to remove the existing reference.

Test Plan:

:eyes: