[Docs/Shared vs Owned] Reword and provide detailed example
Some of the points made in the shared vs owned objects document were
not quite right:

> owned objects are not subject to contention under load

This isn't really true -- if the owned object were hot for some
reason, it is still subject to contension, it just won't be sequenced
by the chain, it will have to be handled off-chain, but someone still
needs to do the work of sequentialising access to it.

> If you want your transaction to execute quickly and consistently,
> then you should prefer owned objects

The "consistently" bit here is strange -- shared object transactions
will also execute consistently -- otherwise the network would fork.

These points have been corrected and the document has been rephrased
to bring up the relative merits of both shared and owned objects, and
replaces the list of examples, with a single worked example showing
those trade-offs in practice when implementing an escrow service.

Test Plan:

:eyes: