Make reputation score abstract (#14584)
## Description 

Starting a little bit on the consensus handler refactoring, and send out
a small PR first just to make sure everyone is ok with the direction. As
you can see in the PR, we are adding a bunch of types inside
consensus-types as abstract trait types, including the Committee and
ConsensusOutput, and then start implement functions on them for Narwhal.
The consensus handler code then calls these functions.
This PR focuses on abstracting the types needed by the reputation score
updates.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes