[GraphQL/Snapshot] Write the current schema before snapshot test
In the previous set-up, if the snapshot test failed, it would need to
be run a second time, after calling `cargo insta review` to ensure
`schema/current_progress_schema.graphql` was properly updated.

There was nothing prompting you to do this which meant that when you
forgot, the changes to `current_progress_schema.graphql` ended up
being applied in the following PR, not the current one.

This PR changes the behaviour so that writing out the current progress
becomes the first operation in the test (before the assertion),
meaning the test only needs to be run once, and then reviewed.

Test Plan:

```
sui-graphql-rpc$  cargo nextest run -- test_schema_sdl_export
```