[GraphQL] Change Object hierarchy representation (#14752)
## Description

A number of changes to bring all the type representations within the
`Object` hierarchy inline with each other, and make it easier to adopt a
`DataLoader` style interface for all of them (which will be the next
step).

### Change `Object` representation

Every type in this hierarchy (`Object`, `MoveObject`, `MovePackage`,
`Coin, `Stake`) stores the following:

- Their `super_` type, by value (e.g. `Coin` stores `MoveObject` which
stores `Object`).
- The "native" representation (the `sui-type` type).

Additionally, `Object`, as the top of the hierarchy also contains its
`stored` representation, and its `address` (to avoid having to convert
it from a byte array every time).

All upcast operations have been made infallible, and work by returning a
reference to `super_`. This is temporary because after #14649, these
will eventually become a true type hierarchy.

All downcasts are implemented by `try_from` implementations that have
been moved into the module that defines the type being created (e.g.
`MovePackage` implements `TryFrom<&Object>`). These convert from a
reference to prevent unnecessarily copying the value when the conversion
might fail.

These `try_from` implementations have been re-used in both the downcast
and `fetch` operations to avoid code duplication.

This type hierarchy now does not use the `SimpleObject` pattern, but
implements each field in an `#[Object]` `impl` block.

### `Eq`, `PartialEq`

I'm not sure why we scattered these derive clauses around everywhere but
I took them out -- because (a) we don't use them, and (b) now that
objects keep the stored representation, we would need to also propagate
the derive clauses into the indexer codebase to use these.

If this is for testing purposes, we should rely on snapshot/expect
testing instead.

### `Result<...>`

If a command can never fail, I've removed the `Result` type, and in code
I have touched, I have made sure we are returning a
`crate::error::Error` and not an `async_graphql::Error` (because the
latter does not apply our error code extension).

I will follow up with an audit across the codebase to make sure this is
done consistently.

### `MovePackage`

`MovePackage` serves requests by first reading its package as a
`ParsedMovePackage`, which is the representation used by `PackageCache`.
To do this required exposing a `Package::read` function as well as some
accessors to the resulting `Package` and `Module` data structures.

This de-duplicates the work of understanding the structure of a package,
and allows fast random access to module and struct information.

Today, this is not done efficiently, because each package will be
re-parsed on each request, but we will follow-up to have this leverage
the package cache (the query will first check whether the package exists
in the cache and re-use it if so, and if not, it will push it into the
cache so the read results can be re-used).

#### `moduleConnection`

There were some bugs in this implication that are now fixed:

`count` was calculated as `last.unwrap_or(first.unwrap_or(DEFAULT))`,
which has the effect of making `last` take precedence. According to the
Relay Connection Spec, the number of edges should be equivalent to
`min(first, last)` if both are present.

Similarly, if both `first` and `last` are present, the algorithm
presented in the Relay Connection Spec states that you should take the
first `firt` from the left, and then take the last `last` from the right
(i.e. they layer on top of each other), which the previous
implementation did not do.

Not really a bug, but a behaviour change: if the edges of the
`Connection` are empty, then return `None` instead of the connection.

### `Stake`

This has been changed quite drastically: Its representation now matches
`Coin` in that it only contains its `super_` representation and the Rust
(`sui-types`) representation of its Move value.

The features that are not currently supported directly in GraphQL (stake
status and estimated rewards) have been implemented by making an RPC
call for the JSON-RPC representation of stake, and then working on that.
I have not been able to test this out because it times out, but at least
this means you can access other fields on the stake object without being
subject to the timeout.

This type should be called `StakedSui` as per the draft schema -- I will
make this change in a follow-up.

## Test Plan

This change is a refactor and therefore is behaviour preserving, so no
extra tests were written, but the following tests were conducted:

```
sui-graphql-rpc$ cargo nextest run
```

Package queries:
![Screenshot 2023-11-09 at 11 55 29
AM](https://github.com/MystenLabs/sui/assets/332275/7ea1c061-1591-4198-a15d-8cfc800ab836)

Object queries:
![Screenshot 2023-11-09 at 11 55 31
AM](https://github.com/MystenLabs/sui/assets/332275/1efe59a4-37c0-45a6-9705-1d8cdceae53b)

Object downcasts:
![Screenshot 2023-11-09 at 11 55 50
AM](https://github.com/MystenLabs/sui/assets/332275/09412351-2fdc-4d3e-84bf-66c82df375fc)

Object upcasts:
![Screenshot 2023-11-09 at 12 34 26
PM](https://github.com/MystenLabs/sui/assets/332275/7cf3f079-f150-4279-a247-6cca7f679515)