Data sim 6/x + Graphql testing: add simulacrum store persistence (#14362)
## Description 

Puts Simulacrum store behind a trait which allows implementing other
backing stores, particularly persistent store using Rocksdb. This is
needed for Graphql testing which wants more control over the storage
layer.

My biggest issue with this change is returning clones over references.
This is because we cannot return ref to objects in DB.
I doubt this will be a big issue as most items are small and perf is not
priority in simulator now.

## Test Plan 

Existing

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes