Separate annotated and unannotated Move values (#14464)
## Description 

This change makes a clean type-level separation between runtime and
annotated Move values. It does this by removing `values` in Move core
types and replaces it with `runtime_value` and `annotated_value`
modules.

This also means that there is no two-tier system of value annotation
either (i.e., `WithFields`). A value is now either fully annotated with
its type and field name information, or a runtime value with no
annotations.

This gets rid of a number of complexities and possible errors if the
wrong format of type was passed into a function.

~~This should not affect or have any impact, and should be fully backwards
compatible, but we should do some additional checks to verify this
before landing this PR.~~ THIS HAS BEEN DONE NOW.

This also removes the `ObjectFormat` struct which was an old hold-over,
and no longer used for anything.

This is a relatively high-touch PR across a number of different areas,
so if you're tagged as a reviewer if you could take a look at the parts
of the code where you are the most familiar and make sure it looks good
to you that would be amazing (and please tag other folks too if you
think they should take a look!).

## Test Plan 

Make sure existing tests pass. 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes