[GraphQL][EASY] Stake -> StakedSui
Change type representing Staked SUI to `StakedSui` (from `Stake`).
This matches the naming used in the draft schema, which was in turn
chosen to match the names used in Move code (as opposed to the names
in JSON-RPC where this concept is also called `Stake`).

Also tweaks the draft schema to make the connection types and names
also use this naming scheme.

Test Plan:

GraphiQL:

<TODO: Screenshot>