[move-ide] Changed the plugin name and updated dependencies (#14791)
## Description 

Changed the name of the Sui-focused plugin to previously agreed name.
Also updated dependencies (TS compiler version etc.). No functionality
added or removed.

## Test Plan 

Verified that it works the same as before
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes