Remove shared object locking in transaction manager before sending to execute (#14732)
## Description 

This PR reverts the logic introduced in #10557 where transaction manager
first executes all the transactions that only read a shared object and
then executes the transaction that mutate it (if there is any). This was
due to that child object reading was not properly handled. Now, since we
have MVCC in child objects, we can remove this restriction, and send all
transactions to execute as soon as all input objects become available.

## Test Plan 

Unit test

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes