[graphql][1/n] conditional db for testing and local dev (#14893)
## Description 

This PR is localized to just the graphql side.
Each backend implements GenericQueryBuilder, and then exports a type
alias QueryBuilder to be used in db_data_provider (pending rename?)
Because the function for executing a query has its type defs tied to the
db backend, we also move this out into the corresponding backend.rs
file. So for pg_backend.rs, we define a new trait PgQueryExecutor and
implement it for PgManager (also pending rename.)

Custom sql functions like 'explain' play nicely enough with this
approach - each backend just needs to implement the QueryFragment and
RunQueryDsl portion.


In subsequent PRs:
1. Rename PgManager, instantiate a reader based on config (which will
just fail if not postgres, until the next one is supported)
2. Modify IndexerReader on indexer crate to support different db
backends

## Test Plan 

Refactor, so existing tests should pass. To pass ci, I've set the
default to pg_backend - this should be fine as we don't have any other
backends right now anyways.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes