[GraphQL/Tests] README for running transactional-tests locally
Instructions for how to set-up `postgresql` for the GraphQL E2E tests.
Also updating the instructions for `sui-indexer` (Postgres is called
`postgresql` on homebrew).

Test Plan:

Follow instructions.  Note that when I did this, the digests and IDs
changed (not sure whether that's supposed to happen or not, but CI
will tell us!).