Enoki SDK improvements (#15050)
## Description 

Some minor updates to the Enoki SDK that are needed after attempting to
integrate into the zkSend codebase:
- The session is currently not observable, which means there's no way to
get the users' JWT consistently in React. `getSession` _mostly_ works
but there is a timing issue if you call it before the `authCallback` has
processed. I updated this state to be observable (the API here is not
really ideal, but we kind of know that _all_ of theses APIs are weird
and will need to be rethought once we have more usage).
- `useAuthCallback` didn't expose the returned state parameter, which is
often used to track auth redirects. The hook now returns the state
parameter so that this functionality can be implemented.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes