[zkLogin] Add length checks (#15057)
## Description 

Revert easy-to-implement + important checks from #14298. Rest requires a
custom json parser (which requires finding an alternative to
https://github.com/microsoft/node-jsonc-parser/) + the remaining are not
as crucial.

## Test Plan 

Added tests!

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes