[graphql/rpc] split out graphiql server (#15064)
## Description 

As mentioned in
[this](https://github.com/MystenLabs/sui/pull/14960#issuecomment-1820785396)
comment:
this is the first part of modularizing the server.
The GraphiQL server with the IDE is decoupled from the core server
Subseq PR will improve the initializers to make setting IDE title driven
by caller at builder time.

The changes here should be opaque to users. No need for release  notes.

## Test Plan 

Existing

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes