[Bridge 3/n] add more queryability to SuiClient (#15144)
## Description 

1. split Event related things to event.rs. Added a macro for more
convenient and SuiEvent declaration
2. implemented `query_events_by_module` and
`get_bridge_events_by_tx_digest` and for SuiClient. The former is a
wrapper impl around the native event query as it paginates by
transaction granularity. In native event query, if a tx has multiple
events, they may be split into different pages, here the implementation
makes sure the last event in the page is the last interesting event in
that transaction. This makes the downstream processing easier by
allowing it to track by tx only.
3. used a trait and added SuiMockClient for testing.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
[Bridge 3/n] add more queryability to SuiClient