[GraphQL/TransactionBlock] ChangeEpochTransaction
Complete support for the ChangeEpoch transaction kind, leveraging the
support for representing `Object`s that have not been indexed to
represent the contents of system package upgrades.

This PR adds the following fields:

- `ChangeEpochTransaction.protocolVersion`.
- `ChangeEpochTransaction.systemPackageConnection`.

And renames `ChangeEpochTransaction.timestamp` to
`ChangeEpochTransaction.startTimestamp` to match its representation in
`sui-types`.

It also uses the native representation of `ChangeEpochTransaction` as
the underlying representation of the GraphQL type, moving it into its
own module.

Test Plan:

It's not easy to simulate a system package upgrade using transactional
tests, so instead, tested manually, by viewing the following
transaction: `F8GGcAVhtPumDoXzc6z9N8Hf9zoTNFx1BPiCy7FJNZsQ` (which is
a change epoch transaction involving a system package):

<TODO: Screenshot>