Add txn and effects as json in analytics (#15151)
## Description 

Given the powerful json querying supported in analytics, we can benefit
a lot from storing these structures as json. Also, we cannot extract
every single detail out of these structs as columns without knowing what
might be useful or needed in the future.
For example - we recently wanted to look at user signatures for a
transaction but it wasn't being captured in any field. If we have these
as json, we can query anything in the future
## Test Plan 

Existing tests