[GraphQL/TransactionBlock] RandomnessStateUpdate (#15185)
Add structured support for the `RandomnessStateUpdate` transaction.

Added a new E2E test that sets up the randomness object, and updates it:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- random.move
```

This required making some changes and additions to the transactional
test runner:

- Epoch changes using `EndOfEpochTransaction` (the newer form) instead
of `ChangeEpoch`.
- A `--create-random-state` option for `advance-epoch`.
- A `set-random-state` command for running the associated system
transaction.
- An API for extracting the latest Epoch ID.

This has knock-on effects to other tests:

- Digests change because the epoch change transaction kind has changed.
- The test for `ChangeEpochTransaction` no longer works (but this will
be fixed once `EndOfEpochTransaction` gains full support in an upcoming
PR).
- (I replaced some fully-qualified types/functions with a `use`).

- #15095
- #15145
- #15146
- #15147
- #15179
- #15180
- #15181
- #15182
- #15183
- #15184