[GraphQL/MoveObject] Remove `hasPublicTransfer`
Since we made it possible to add abilities to existing system types
through protocol upgrades, this field is no longer accurate:
Previously, when an object was created, this value was cached based on
whether its type had `store` at the time that version of the object
was created.

This can cause problems when it's a system type, and `store` was added
to it in a protocol upgrade, as the field will incorrectly state that
the object does not have public transfer when in fact it does.

This was fixed at the protocol level by recomputing the field every
time the object is loaded for execution, but it still means the field
exists and could be stale.

This PR removes the field from GraphQL to avoid confusion due to
staleness.

Test Plan:

```
sui-graphql-rpc$ cargo nextest run
```