[GraphQL/Schema] Merge System State into Epoch (#15218)
## Description

Renames `SuiSystemStateSummary` to `SystemState` and makes it a field of
`Epoch` based on discussion in our latest weekly meeting. If
`Query.latestSystemStateSummary` was part of the draft schema, I would
remove that too, but it's already not there. The recommended way to get
the latest system state would be to do:

```graphql
{
  epoch {
    systemState {
      # ...
    }
  }
}
```

This PR also deduplicates some fields that were previously in both
`Epoch` and `SuiSystemStateSummary`: (`protocolConfigs`,
`referenceGasPrice` and `validatorSet`).

## Test Plan

:eyes: