[GraphQL/TransactionBlock] ProgrammableTransactionBlock (#15211)
## Description

Structured representation of ProgrammableTransactionBlock. The
implementation diverges from the originally written schema in some
important ways:

- `TransactionInput` has changed drastically. Previously it was a union
between `MoveObject`, `MovePackage` and `SharedInput`. Now it is a union
between `ImmutableOrOwned` (object), `SharedInput`, `Receiving`, and
`Pure`. This is because it's not possible to know from the transaction
kind alone whether a given input is a package or an object, so they have
been combined into `ImmutableOrOwned`, the absence of `Pure` arguments
was a previous oversight, and `Receiving` is a new feature.

- Types that were empty in the schema now need a dummy field.

- The `inputs` and `transactions` fields are now connections, because
they can contain complex sub-queries that will need to be limited.

- `ImmutableOrOwned` transaction inputs expose the object being
referenced, but also its object reference. This is so the object being
referenced can be identified even if the indexer does not have its
contents (e.g. it's pruned).

- `MoveCallTransaction` similarly exposes the details of the function
being called (package, module, name) as well as the resolved function,
so that even for transactions that failed to run because the function
could not be found, we can display which function it was.

These changes will be reflected back into the draft schema at a later
date, along with other smaller discrepencies that have crept in during
development.

## Test Plan

Updated the PTB E2E test:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- programmable.move
```

This updated test required increasing the query node limit in test
settings. Something to consider when revisiting the limits.

## Stack

- #15095 
- #15145 
- #15146 
- #15147
- #15179
- #15180 
- #15181 
- #15182 
- #15183 
- #15184 
- #15185 
- #15186 
- #15188