[move 2024] Combine namespaces for expansion lookup (#14978)
## Description 

This adds a Move 2024 path resolution mechanism to expansion, which
revises name chain resolution to combine address, module, and member
namespaces. In addition, it adds `::<addr>::<module>` and
`::<addr>::<module>::<member>` forms to alias-dodge when necessary.

This is toward getting ready to add enums to the name chain space.

## Test Plan 

New and updated tests, subject to feature gating.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes