[Docs] add zkLogin example (#15123)
## Description 

In order to facilitate a comprehensive understanding of each step
involved in Sui zkLogin for developers, I crafted a zkLogin Example. (by
the way, do I need to sign my name on the this docs?

I'm unsure if this aligns with the specifications. If any modifications
are needed, please feel free to reach out, and I would be happy to make
the necessary adjustments!

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: hayes-mysten <135670682+hayes-mysten@users.noreply.github.com>
Co-authored-by: William Robertson <williamrobertson@mystenlabs.com>
Co-authored-by: ronny-mysten <118224482+ronny-mysten@users.noreply.github.com>