[snapshots] Upload formal snapshots before db snapshots (#15261)
Formal snapshots are much smaller than db snapshots, hence they take
much less time to upload. Additionally, validators are the greatest
beneficiaries of formal snapshots, and are more sensitive to the effects
of lagging the network, therefore there is a strong need for formal
snapshots of the current epoch to be available as quickly as possible
after epoch change (6 hour lag out of a 24 hour epoch is too long).

To address this, for a node that is configured to upload both snapshot
types, formal snapshots will be the first to start, and db snapshot
upload will block on its completion (note that we cannot do them
concurrently as they both need to open the db checkpoint as a perpetual
DB - formal snapshots in order to scan the live object set, and db
checkpoint uploader to prune and compact.

Ran locally and verified snapshot and db checkpoints are being generated
successfully

```
williamsmith in /opt/sui/db/authorities_db/full_node_db λ du -sh ./*
  0B	./db_checkpoints
 51G	./live
  0B	./snapshot
williamsmith in /opt/sui/db λ du -sh ./local_*
 76G	./local_db_snapshots
 26G	./local_formal_snapshots
```

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration