[cli] Validate alias (#15307)
## Description 

Introduce a validation step when creating or updating an alias. The
alias must start with a letter and can contain only letters, digits,
hyphens (-), or underscores (_).

## Test Plan 

Updated existing tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Added a validation step for alias naming when creating or updating the alias of an address
in the Sui CLI. The alias must start with a letter and can contain only
letters, digits, hyphens (-), or underscores (_).