[bridge 8/n] make authority aggregation function more generic (#15290)
## Description 


`BridgeClient` needs to talk to bridge authorities similar to how
AuthorityAggregator works. This PR make
`quorum_map_then_reduce_with_timeout_and_prefs` and
`quorum_map_then_reduce_with_timeout` more generic, and move them to a
new crate `sui-common` because we don't want `sui-bridge` to depend on
`sui-core`.

Why it's not in `mysten-common` crate? because that crate does not
depend on any `sui` stuff and i don't want to break it..

The most important thing in this PR is to make sure the original member
functions now trait method `weight` and `shuffle_by_stake_with_rng` does
not change.



## Test Plan 

All existing tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
[bridge 8/n] make authority aggregation function more generic