[GraphQL/Package Resolver] Limits on type parameter nesting
Limit how deeply type parameters can be nested in a type tag that
needs to be resolved.  This limit only applies to the "outermost" type
parameter application, and not to type parameters that exist in
signatures in Move bytecode (which already had to go through
chain-enforced limits to be published).

Test Plan:

New unit tests:

```
sui-package-resolver$ cargo nextest run
```