Start writing manifest file in formal and db snapshot buckets (#15389)
## Description 

We do not have a way to find out the latest snapshot available to
download in different snapshot buckets. While it would be possible to
just list the epoch dirs in a bucket, it is not possible to do so with
passwordless. This PR adds a background task which uploads a MANIFEST
file in snapshot buckets which has a list of available snapshots at any
given point in time.
We will leverage this in sui-tool to download latest snapshot and a new
endpoint which lists all available snapshot

## Test Plan 

Ran locally