[bridge 11/n] handle eth actions in orchestrator (#15336)
## Description 

1. For Orchestrator's eth watcher channel, change eth logs from `Log` to
`EthLog`. The latter contains essential attribute of the log, which are
Optional in the former. See `fn get_log_tx_details` for the conversion
from `Log` to `EthLog`
2. add handling for eth actions (handling for sui actions is done in
previous PR)
3. add storage focused tests for orchestrator


## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes